requirements.txt
#lista de paquetes a instalar que no vienen por defecto
pandas>=1.0.3
requests>=2.27.1
psycopg2-binary>=2.9.6
flask>=2.1.3
gunicorn==20.1.0
